Collapse widget after click for nicer UI behaviour
This hides the menu after clicking on one of the actions, so it doesn't need to be dismissed manually when the machine comes back from suspend/lockscreen etc.
Collapsing the menu also makes sense for rebooting/shutdown/logging out, because sometimes it can take some time before the actual session termination takes place. The old behaviour in this case was that the menu stays visible after click, which could cause some doubt if the click was actually registered by the widget.